extract build order iterator factory, renaming
